pyproject.toml
[project]
name = "open-data1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.75.0",
    "duckdb>=1.4.3",
    "google-generativeai>=0.8.5",
    "httpx>=0.28.1",
    "jupyter>=1.1.1",
    "litellm>=1.80.10",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pyarrow>=22.0.0",
    "pydantic>=2.12.5",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "scipy>=1.16.3",
    "tenacity>=9.1.2",
    "tqdm>=4.67.1",
]
